undo some casts

+++ src/symbol.c
@@ -102,7 +102,7 @@ static void lay_out_union(struct global_symbols_t *S, struct symbol *sym, struct
 			info->max_align = sym->ctype.alignment;
 	}
 
-	if (sym->bit_size > (int)info->bit_size)
+	if (sym->bit_size > info->bit_size)
 		info->bit_size = sym->bit_size;
 
 	sym->offset = 0;
@@ -235,7 +235,7 @@ static struct symbol * examine_array_type(struct global_symbols_t *S, struct sym
 
 	if (array_size) {	
 		bit_size = array_element_offset(S->C->target, base_type->bit_size,
-						(int) get_expression_value_silent(S->C, array_size));
+						get_expression_value_silent(S->C, array_size));
 		if (array_size->type != EXPR_VALUE) {
 			if (S->C->Wvla)
 				warning(S->C, array_size->pos, "Variable length array is used.");
@@ -257,7 +257,7 @@ static struct symbol *examine_bitfield_type(struct global_symbols_t *S, struct s
 	if (!base_type)
 		return sym;
 	bit_size = base_type->bit_size;
-	if (sym->bit_size > (int)bit_size)
+	if (sym->bit_size > bit_size)
 		warning(S->C, sym->pos, "impossible field-width, %d, for this type",  sym->bit_size);
 
 	alignment = base_type->ctype.alignment;
@@ -309,7 +309,7 @@ static int count_array_initializer(struct global_symbols_t *S, struct symbol *t,
 			count++;
 			switch (entry->type) {
 			case EXPR_INDEX:
-				if ((int)entry->idx_to >= nr)
+				if (entry->idx_to >= nr)
 					nr = entry->idx_to+1;
 				break;
 			case EXPR_PREOP: {
